requirements.txt
amqp==5.1.1
asgiref==3.5.2
astroid==2.12.9
billiard==3.6.4.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
defusedxml==0.7.1
dill==0.3.5.1
Django==4.1.1
django-allauth==0.51.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==3.3
isort==5.10.1
kombu==5.2.4
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.1
Pillow==9.2.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pylint==2.15.2
pylint-celery==0.3
pylint-django==2.5.3
pylint-plugin-utils==0.7
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
tomli==2.0.1
tomlkit==0.11.4
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
